requirements.txt
python-telegram-bot>=22.2,<23
python-dotenv>=1.0